Can't remove last fornece_sec for a produto. Removing a produto now causes the removal of all correspondent lines in fornece_sec (already did), planograma and reposicao.

+++ web/remove_forn_secundario.php
@@ -15,16 +15,27 @@
                 $user ="ist180832";
                 $password = "carreiracarreira";
                 $dbname = $user;
-            
                 $db = new PDO("pgsql:host=$host;dbname=$dbname", $user, $password);
                 $db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
+                $sql = "SELECT COUNT(nif)
+                        FROM fornece_sec
+                        WHERE ean = $ean;";
+                $result = $db->query($sql);
+                $count = $result->fetchColumn();
+                
+                if($count == 1) {
+                    echo "<p>Não é possível remover o último fornecedor secundário de um produto.</p>";
+                    echo "<p>Por favor adicione outro fornecedor secundário a <b>$design</b> antes de remover o fornecedor nif = <b>$forn_sec</b></p>";
+                }
 
-                $sql = "DELETE FROM fornece_sec WHERE nif='$forn_sec' AND ean='$ean';";
-
-                echo("<p>Removendo o fornecedor secundário '$forn_sec' do produto (EAN = $ean, designação = '$design'):</p>");
-                echo($sql);
-                $db->query($sql);
+                else {
+                    $sql = "DELETE FROM fornece_sec WHERE nif='$forn_sec' AND ean='$ean';";
+                
+                    echo("<p>Removendo o fornecedor secundário '$forn_sec' do produto (EAN = $ean, designação = '$design'):</p>");
+                    echo($sql);
+                    $db->query($sql);
+                }
 
                 $db = null;
             }
@@ -33,6 +44,7 @@
                 echo("<p>ERROR: {$e->getMessage()}</p>");
             }
             echo("<p><a href=\"ver_fornecedores.php?ean=$ean&design=$design&forn_prim=$forn_prim\"> Voltar</a></p>");
+            echo("<p><a href=\"supermercado.php\">Ver supermercado</a></p>");
 
         ?>
     </body>
